requirements.txt
requests~=2.32.3
python-dotenv~=1.0.1
beautifulsoup4~=4.12.3
lxml~=5.2.1
langgraph~=0.2.76
langchain-together~=0.3.0
newspaper3k~=0.2.8
lxml-html-clean~=0.4.2
streamlit~=1.45.1
prefect~=3.4.1
langchain-nvidia-ai-endpoints~=0.3.9
langgraph~=0.2.76
exa_py~=1.9.1
langchain~=0.3.25
fastapi
uvicorn
